Refactor Fluctuation into non-Savings Namespace
